requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0  # EB推荐的WSGI服务器
whitenoise>=6.5.0  # 静态文件服务
